editing project mostly done, images and categories left, god help us all

+++ client/src/components/Admin/Dashboard/ProjectsView/ProjectsView.js
@@ -1,7 +1,8 @@
 import React, { useState, useEffect } from 'react';
 import './ProjectsView.css';
-import { FaTrash, FaMapPin, FaSearch, FaArrowLeft } from 'react-icons/fa';
+import { FaTrash, FaMapPin, FaSearch, FaArrowLeft, FaPlus } from 'react-icons/fa';
 import ProjectAdminDetails from './ProjectAdminDetails/ProjectAdminDetails';
+import swal from 'sweetalert';
 
 function ProjectsView(props){
     const
@@ -30,19 +31,28 @@ function ProjectsView(props){
        'Architecture'
     ];
 
-   const deleteProject = ((_id) => {
-    if(window.confirm('Are you sure you want to delete this project?'))
-    fetch(`${api}/projects`,  {
-        method: 'delete',
-        headers: {'Content-Type': 'application/json', token},
-        body: JSON.stringify({_id})
-      })
-      .then(res => res.json())
-      .then(res => {
-        fetch(`${api}/projects`)
-        .then(res => res.json())
-        .then(res => setProjects(res));
-      });
+   const deleteProject = ((_id, title) => {
+    swal({
+        title: `Deleting ${title}`,
+        text: `Are you sure that you want to delete this project?`,
+        icon: "warning",
+        dangerMode: true,
+        buttons: {
+            cancel: true,
+            confirm: true,
+        },
+    })
+    .then(willDelete => {
+        if(willDelete){
+            fetch(`${api}/projects`,  {
+                method: 'delete',
+                headers: {'Content-Type': 'application/json', token},
+                body: JSON.stringify({_id})
+            })
+            .then(res => res.json())
+            .then(res => refreshProjects());
+        }
+    })
    });
 
    const refreshProjects = () => {
@@ -66,6 +76,7 @@ function ProjectsView(props){
     return((!projects.length) ? <div>Loading...</div> :
         (!viewingProject) ? 
         <div>
+            <div className="new-project"><FaPlus /></div>
             <div>
             <div className="projects-search">
                 <div className="projects-search-div">
@@ -77,25 +88,21 @@ function ProjectsView(props){
             <div className="viewed-projects-container">
                 {viewedProjects.map((element) => {
                     let description = element.description.en;
-                if (description.length > 100)
-                    description = description.substr(0, 150) + '...';
+                if (description.length > 200)
+                    description = description.substr(0, 200) + '...';
                     return(
-                        <div onClick={() => {setViewingProject(element)}} className="viewed-project" key={element._id}>
-                            <div className="preview-div">
+                        <div className="viewed-project" key={element._id}>
+                            <div onClick={() => {setViewingProject(element)}} className="preview-div">
                                 <img src={element.preview} />
                             </div>
-                            <div className="viewed-project-div">
+                            <div onClick={() => {setViewingProject(element)}} className="viewed-project-div">
                                 <span className="viewed-project-title">{element.title.en}</span>
                                 <span className="viewed-project-sub-title"><FaMapPin /> {element.location.en}</span>
                                 <p className="viewed-project-subtitle">{categories[element.category]}, {subcategories[element.subcategory]}</p>
                                 <p className="viewed-project-description">{description}</p>
                             </div>
                             <div className="viewed-project-delete-div">
-                                {/* <div className="viewed-project-delete" onClick={() => deleteProject(element._id)}>
-                                    <div className="viewed-project-edit-icon"><FaEdit /></div>
-                                    <div className="viewed-project-edit-text">Edit</div>
-                                </div> */}
-                                <div className="viewed-project-delete" onClick={() => deleteProject(element._id)}>
+                                <div className="viewed-project-delete" onClick={() => deleteProject(element._id, element.title.en)}>
                                     <div className="viewed-project-delete-icon"><FaTrash /></div>
                                     <div className="viewed-project-delete-text">Delete</div>
                                 </div>
